+++ cmd/example.go
@@ -1,8 +1,7 @@
 package main
 
 import (
-	"github.com/c0x12c/numerator-go-sdk/pkg/l"
-	"log"
+	"github.com/c0x12c/numerator-go-sdk/pkg/log"
 
 	"github.com/c0x12c/numerator-go-sdk/internal/clients"
 	"github.com/c0x12c/numerator-go-sdk/pkg/config"
@@ -19,11 +18,8 @@ func main() {
 	apiKey := viper.Get("API_KEY").(string)
 	numeratorConfig := config.NewNumeratorConfig(apiKey)
 
-	// Create a l instance
-	logger, err := l.NewZapLogger()
-	if err != nil {
-		log.Fatalf("failed to create l: %v", err)
-	}
+	// Create a log instance
+	logger, _ := log.NewZapLogger()
 
 	// Create Numerator client
 	numeratorClient := clients.NewDefaultNumeratorClient(numeratorConfig)
@@ -31,19 +27,19 @@ func main() {
 	// Fetch feature flags
 	flags, err := numeratorClient.FeatureFlags(constant.Page, constant.Size)
 	if err != nil {
-		logger.Error("failed to fetch feature flags", l.Error(err))
+		logger.Error("failed to fetch feature flags", log.Error(err))
 		return
 	}
-	logger.Info("fetched feature flags", l.Any("flags", flags))
+	logger.Info("fetched feature flags", log.Any("flags", flags))
 
 	// Fetch feature flag details
 	flagKey := "go_featureflag_01"
 	flagDetail, err := numeratorClient.FeatureFlagDetails(flagKey)
 	if err != nil {
-		logger.Error("failed to fetch feature flag details", l.Error(err))
+		logger.Error("failed to fetch feature flag details", log.Error(err))
 		return
 	}
-	logger.Info("fetched feature flag details", l.Any("flagDetail", flagDetail))
+	logger.Info("fetched feature flag details", log.Any("flagDetail", flagDetail))
 
 	// Create an empty context
 	context := make(map[string]interface{})
@@ -52,7 +48,7 @@ func main() {
 	defaultBoolean := true
 	booleanValue, err := numeratorClient.GetValueByKeyWithDefault(flagKey, context, defaultBoolean)
 	if err != nil {
-		logger.Error("failed to fetch flag value by key", l.Error(err))
+		logger.Error("failed to fetch flag value by key", log.Error(err))
 		return
 	}
 	// Perform type assertion to ensure booleanValue is a boolean
@@ -61,7 +57,7 @@ func main() {
 		logger.Error("failed to fetch flag value by key: booleanValue is not a boolean")
 		return
 	}
-	logger.Info("fetched flag value by key", l.Any("BooleanValue", boolVal))
+	logger.Info("fetched flag value by key", log.Any("BooleanValue", boolVal))
 
 	// Create a context
 	contextEnv := map[string]interface{}{
@@ -71,7 +67,7 @@ func main() {
 	// Fetch feature flag value by key with empty context
 	booleanValue, err = numeratorClient.GetValueByKeyWithDefault(flagKey, contextEnv, defaultBoolean)
 	if err != nil {
-		logger.Error("failed to fetch flag value by key", l.Error(err))
+		logger.Error("failed to fetch flag value by key", log.Error(err))
 		return
 	}
 	// Perform type assertion to ensure booleanValue is a boolean
@@ -80,5 +76,5 @@ func main() {
 		logger.Error("failed to fetch flag value by key: booleanValue is not a boolean")
 		return
 	}
-	logger.Info("fetched flag value by key with context", l.Any("BooleanValue", boolVal))
+	logger.Info("fetched flag value by key with context", log.Any("BooleanValue", boolVal))
 }
