+++ pkg/log/logger.go
@@ -1,4 +1,4 @@
-package l
+package log
 
 type Logger interface {
 	Info(message string, args ...LogArgs)
@@ -9,4 +9,18 @@ type Logger interface {
 type LogArgs struct {
 	Key   string
 	Value any
+}
+
+func Any(key string, value any) LogArgs {
+	return LogArgs{
+		Key:   key,
+		Value: value,
+	}
+}
+
+func Error(err error) LogArgs {
+	return LogArgs{
+		Key:   "error",
+		Value: err,
+	}
 }
